Plot the oral equivalent dose (OED) and exposure data (if applicable)

This function plots all OEDs and any uploaded exposure data. How the OEDs are plotted is dependent upon the user-selected output type (single OED quantile or all OED samples). Exposure data (if applicable) is plotted as a point range.

Usage

IVIVEplotting(OED_data, BioactiveConc, pars, logscale, expdata)

Arguments

OED_data

A data frame of chemical names and their OEDs (either one per chemical or many per chemical if samples were returned); The first output from the IVIVEsol function

BioactiveConc

An input not used

pars

A list of all user input parameters for the entire app

logscale

Checkbox input value indicating if the user wanted the y-axis of plots to be a log10 scale

expdata

A data frame with chemical names and exposure data estimates

Value

A ggplot2 plotting object

See also