Are there any required hepatic clearance (Clint) or fraction unbound in plasma (fup) values missing from the compounds file uploaded by the user?

This function checks if the uploaded file with physical-chemical data has any missing hepatic clearance or fraction unbound in plasma values. If so, a message is returned to the user.

Usage

MissingClintFup(file, spec, defaultHuman)

Arguments

file

The data frame with compound physical-chemical data uploaded by the user

spec

The species selected by the user; either "Dog","Human","Mouse", "Rabbit", "Rat"

defaultHuman

The default to human status selected by the user; either "Yes" if the user wants to use human in vitro data in place of missing animal in vitro data or "No" to only use animal in vitro data

Value

If any required data is missing, then this function returns a message detailing the missing data. If all required data is present, then no message is outputted.

See also