Generates a drop down list of possible models to simulate

Generates a drop down list of possible models to simulate

Usage

Model_Input(func, spec)

Arguments

func

The user-selected output

spec

The user-selected species

Value

A drop down list

See also

Examples

Model_Input("Concentration-time profiles","Human")
#> <div class="form-group shiny-input-container">
#>   <label class="control-label" id="model-label" for="model">Select a model to simulate.</label>
#>   <div>
#>     <select id="model" class="shiny-input-select"><option value="Select" selected>Select</option>
#> <option value="1compartment">1compartment</option>
#> <option value="3compartment">3compartment</option>
#> <option value="pbtk">pbtk</option>
#> <option value="fetal_pbtk">fetal_pbtk</option></select>
#>     <script type="application/json" data-for="model" data-nonempty="">{"plugins":["selectize-plugin-a11y"]}</script>
#>   </div>
#> </div>
Model_Input("Parameter calculations","Rat")
#> <div class="form-group shiny-input-container">
#>   <label class="control-label" id="model-label" for="model">Select a model to simulate.</label>
#>   <div>
#>     <select id="model" class="shiny-input-select"><option value="Select" selected>Select</option>
#> <option value="Schmitt">Schmitt</option></select>
#>     <script type="application/json" data-for="model" data-nonempty="">{"plugins":["selectize-plugin-a11y"]}</script>
#>   </div>
#> </div>