Compile two outputs for the 'Preloaded Compounds' card under the 'Compound Selection' card

This function presents the user interface for the preloaded compounds section of the 'Preloaded Compounds' card under the 'Compound Selection' tab

Usage

PreloadComps_UI(func, spec, defaulthuman, model, insilicopars, honda, comptype)

Arguments

func

User selection with the input$func ID; either "Concentration-time profiles", "Steady state concentrations", "In vitro in vivo extrapolation (IVIVE)", and "Parameter calculations"

spec

User selection with the input$spec ID; either "Dog","Human","Mouse", "Rabbit","Rat"

defaulthuman

User selection with input$defaulttoHuman; either "Yes" for using human in vitro data in place of animal data or "No" for only using animal data

model

User selection with the input$model ID; either "3compartmentss", "Schmitt","1compartment","3compartment", "pbtk" or "fetal_pbtk"

insilicopars

User selection with the input$insilicopars ID; either "Yes, load in silico parameters" or "No, do not load in silico parameters"

honda

User selection with the input$HondaIVIVE ID; either "NULL","Honda1", "Honda2","Honda3","Honda4"

comptype

User selection with the input$comptype ID; either "Choose from all available chemicals" or "Choose from only food relevant chemicals"

Value

The user interface objects (drop down selections) for the 'Preloaded Compounds' card on the 'Compound Selection' tab. The objects outputted vary depending on the user's input selection.

See also