Remove current plot legend and generate individual plot legends

This function extracts the plotting legend from the individ_plt_lst, creates a new plotting object with just the legend, and then removes all the legends from the remaining plot.

Usage

Set_Individ_Plot(individ_plt_lst, n_states)

Arguments

individ_plt_lst

A list of ggplot2 plotting objects where each list element is the plot of a single compound and is composed of subplots

n_states

Number of model compartments/outputs

Value

A list containing the inputted plot list with the legends on each compound plot removed and a new subplot with the plotting legend

See also