Create the text for a figure/table caption

This function the correct AUC output concentration to go in a caption in the concentration-time profiles module plot and table.

Usage

caption_text(func, model)

Arguments

func

User selection with the input$func ID; either "Concentration-time profiles", "Steady state concentrations", "In vitro in vivo extrapolation (IVIVE)", and "Parameter calculations"

model

User selection with the input$model ID; either "3compartmentss", "Schmitt","1compartment","3compartment", "pbtk" or "fetal_pbtk"

Value

A string with text describing the returned concentration of the model.

See also