Generate the user upload option to upload compounds and their physical-chemical data

Generate the user upload option to upload compounds and their physical-chemical data

Usage

fileInput_UploadedComps(id)

Arguments

id

Shiny identifier name

Value

A browser and file upload box

See also

Examples

fileInput_UploadedComps("file1")
#> <div class="form-group shiny-input-container">
#>   <label class="control-label" id="file1-label" for="file1">Upload a CSV file of physical and chemical data for compounds not
#>                     in the preloaded list (if desired). Download the 'Uploaded
#>                     Physical-Chemical Data File Folder' under the 'Instructions' card
#>                     for file formatting instructions.</label>
#>   <div class="input-group">
#>     <label class="input-group-btn input-group-prepend">
#>       <span class="btn btn-default btn-file">
#>         Browse...
#>         <input id="file1" class="shiny-input-file" name="file1" type="file" style="position: absolute !important; top: -99999px !important; left: -99999px !important;" accept=".csv"/>
#>       </span>
#>     </label>
#>     <input type="text" class="form-control" placeholder="No file selected" readonly="readonly"/>
#>   </div>
#>   <div id="file1_progress" class="progress active shiny-file-input-progress">
#>     <div class="progress-bar"></div>
#>   </div>
#> </div>