Generate the ggplot2 plotting object for the multi-curve plot

This function creates plots (number of plots = number of model compartments outputted from httk's solve_model function) that display the concentration- time profile curves for all simulated compounds within each model compartment. Plots are created using ggplot from the ggplot2 R package. A list of plots are returned (one plot per compartment plus a plot displaying the plot legend).

Usage

plottingfunc_all(sol_array)

Arguments

sol_array

The concentration-time profile solution array where entries signify the concentration of a compound at a specific time in a particular compartment. Pages of the array correspond to compounds, rows to time points, and columns to model compartments and outputs.

Value

A list containing two elements, the first one being a plot list with each element of the list representing a different compound (no plot legends) and the second element being a data frame of hexadecimal RGB triplet colors that defines the plotting color for each compound

See also