Generate the plots for the individual plots drop down

This function creates N plots, where N is the number of compounds simulated in the concentration-time profile simulations. The plot list for each compound consists of an individual plotting object for each model compartment.

Usage

plottingfunc_individual(sol_array, plt_colors)

Arguments

sol_array

The concentration-time profile solution array where entries signify the concentration of a compound at a specific time in a particular compartment. Pages of the array correspond to compounds, rows to time points, and columns to model compartments and outputs.

plt_colors

A data frame of colors where each entry is a hexadecimal RGB triplet for one compound

Value

A list that contains other lists. Each list element makes up one full plotting figure for one compound (i.e. a plot with subplots).

See also