Generates a drop down list of tissue types for the steady state concentration and IVIVE outputs

Generates a drop down list of tissue types for the steady state concentration and IVIVE outputs

Usage

selectInput_Tissue(id)

Arguments

id

Shiny identifier

Value

A drop down list

See also

Examples

selectInput_Tissue("tissueSS")
#> <div class="form-group shiny-input-container">
#>   <label class="control-label" id="tissueSS-label" for="tissueSS">Select a tissue you want the output concentration in.
#>                       Leave on 'NULL' if the whole body concentration is desired.</label>
#>   <div>
#>     <select id="tissueSS" class="shiny-input-select"><option value="NULL" selected>NULL</option>
#> <option value="adipose">adipose</option>
#> <option value="bone">bone</option>
#> <option value="brain">brain</option>
#> <option value="gut">gut</option>
#> <option value="heart">heart</option>
#> <option value="kidney">kidney</option>
#> <option value="liver">liver</option>
#> <option value="lung">lung</option>
#> <option value="muscle">muscle</option>
#> <option value="skin">skin</option>
#> <option value="spleen">spleen</option>
#> <option value="rest">rest</option></select>
#>     <script type="application/json" data-for="tissueSS" data-nonempty="">{"plugins":["selectize-plugin-a11y"]}</script>
#>   </div>
#> </div>
selectInput_Tissue("tissueIVIVE")
#> <div class="form-group shiny-input-container">
#>   <label class="control-label" id="tissueIVIVE-label" for="tissueIVIVE">Select a tissue you want the output concentration in.
#>                       Leave on 'NULL' if the whole body concentration is desired.</label>
#>   <div>
#>     <select id="tissueIVIVE" class="shiny-input-select"><option value="NULL" selected>NULL</option>
#> <option value="adipose">adipose</option>
#> <option value="bone">bone</option>
#> <option value="brain">brain</option>
#> <option value="gut">gut</option>
#> <option value="heart">heart</option>
#> <option value="kidney">kidney</option>
#> <option value="liver">liver</option>
#> <option value="lung">lung</option>
#> <option value="muscle">muscle</option>
#> <option value="skin">skin</option>
#> <option value="spleen">spleen</option>
#> <option value="rest">rest</option></select>
#>     <script type="application/json" data-for="tissueIVIVE" data-nonempty="">{"plugins":["selectize-plugin-a11y"]}</script>
#>   </div>
#> </div>